+++ cartridge/failover.lua
@@ -26,12 +26,14 @@
 -- @local
 
 local log = require('log')
+local json = require('json')
 local fiber = require('fiber')
 local checks = require('checks')
 local errors = require('errors')
 local membership = require('membership')
 
 local vars = require('cartridge.vars').new('cartridge.failover')
+local pool = require('cartridge.pool')
 local utils = require('cartridge.utils')
 local topology = require('cartridge.topology')
 local service_registry = require('cartridge.service-registry')
@@ -39,11 +41,13 @@ local stateboard_client = require('cartridge.stateboard-client')
 local etcd2_client = require('cartridge.etcd2-client')
 
 local FailoverError = errors.new_class('FailoverError')
+local SwitchoverError = errors.new_class('SwitchoverError')
 local ApplyConfigError = errors.new_class('ApplyConfigError')
 local ValidateConfigError = errors.new_class('ValidateConfigError')
 local StateProviderError = errors.new_class('StateProviderError')
 
 vars:new('membership_notification', membership.subscribe())
+vars:new('consistency_needed', false)
 vars:new('clusterwide_config')
 vars:new('failover_fiber')
 vars:new('failover_err')
@@ -51,13 +55,24 @@ vars:new('schedule', {})
 vars:new('client')
 vars:new('cache', {
     active_leaders = {--[[ [replicaset_uuid] = leader_uuid ]]},
+    is_vclockkeeper = false,
     is_leader = false,
     is_rw = false,
 })
 vars:new('options', {
+    WAITLSN_PAUSE = 0.2,
+    WAITLSN_TIMEOUT = 3,
     LONGPOLL_TIMEOUT = 30,
     NETBOX_CALL_TIMEOUT = 1,
 })
+
+function _G.__cartridge_failover_get_lsn(timeout)
+    box.ctl.wait_ro(timeout)
+    return {
+        id  = box.info.id,
+        lsn = box.info.lsn,
+    }
+end
 
 --- Generate appointments according to clusterwide configuration.
 -- Used in 'disabled' failover mode.
@@ -156,14 +171,6 @@ local function accept_appointments(appointments)
         end
     end
 
-    if active_leaders[box.info.cluster.uuid] == box.info.uuid then
-        vars.cache.is_leader = true
-        vars.cache.is_rw = true
-    else
-        vars.cache.is_leader = false
-        vars.cache.is_rw = replicasets[box.info.cluster.uuid].all_rw
-    end
-
     if utils.deepcmp(vars.cache.active_leaders, active_leaders) then
         return false
     end
@@ -200,9 +207,146 @@ local function apply_config(mod)
     )
 end
 
-local function reconfigure_all()
+local function constitute_oneself(active_leaders, opts)
+    checks('table', {
+        timeout = 'number',
+    })
+
+    local confapplier = require('cartridge.confapplier')
+    local instance_uuid = confapplier.get_instance_uuid()
+    local replicaset_uuid = confapplier.get_replicaset_uuid()
+
+    local topology_cfg = vars.clusterwide_config:get_readonly('topology')
+
+    if active_leaders[replicaset_uuid] ~= instance_uuid then
+        -- I'm not a leader
+        vars.cache.is_vclockkeeper = false
+        vars.cache.is_leader = false
+        vars.cache.is_rw = topology_cfg.replicasets[replicaset_uuid].all_rw
+        return true
+    end
+
+    -- Get ready to become a leader
+    if not vars.consistency_needed then
+        vars.cache.is_vclockkeeper = false
+        vars.cache.is_leader = true
+        vars.cache.is_rw = true
+        return true
+    elseif vars.cache.is_vclockkeeper then
+        -- I'm already a vclockkeeper
+        return true
+    end
+
+    local deadline = fiber.time() + opts.timeout
+
+    -- Go to the external state provider
+    local session = assert(vars.client):get_session()
+    if session == nil or not session:is_alive() then
+        return nil, StateProviderError:new('State provider unavailable')
+    end
+
+    -- Query current vclockkeeper
+    -- WARNING: implicit yield
+    local vclockkeeper, err = session:get_vclockkeeper(replicaset_uuid)
+    fiber.testcancel()
+    if err ~= nil then
+        return nil, SwitchoverError:new(err)
+    end
+
+    if vclockkeeper == nil then
+        -- It's absent, no need to wait anyone
+        goto set_vclockkeeper
+    elseif vclockkeeper.instance_uuid == instance_uuid then
+        -- It's me already, but vclock still should be persisted
+        goto set_vclockkeeper
+    end
+
+    do
+        -- Go to the vclockkeeper and query its LSN
+        local vclockkeeper_uuid = vclockkeeper.instance_uuid
+        local vclockkeeper_srv = topology_cfg.servers[vclockkeeper_uuid]
+        if vclockkeeper_srv == nil then
+            return nil, SwitchoverError:new('Alien vclockkeeper %q',
+                vclockkeeper.instance_uuid
+            )
+        end
+
+        local vclockkeeper_uri = vclockkeeper_srv.uri
+
+        local timeout = deadline - fiber.time()
+        -- WARNING: implicit yield
+        local vclockkeeper_info, err = errors.netbox_call(
+            pool.connect(vclockkeeper_uri, {wait_connected = false}),
+            '__cartridge_failover_get_lsn', {timeout},
+            {timeout = timeout + vars.options.NETBOX_CALL_TIMEOUT}
+        )
+        fiber.testcancel()
+
+        if vclockkeeper_info == nil then
+            return nil, SwitchoverError:new(err)
+        end
+
+        -- Wait async replication to arrive
+        -- WARNING: implicit yield
+        local timeout = deadline - fiber.time()
+        local ok = utils.wait_lsn(
+            vclockkeeper_info.id,
+            vclockkeeper_info.lsn,
+            vars.options.WAITLSN_PAUSE,
+            timeout
+        )
+        fiber.testcancel()
+        if not ok then
+            return nil, SwitchoverError:new(
+                "Can't catch up with the vclockkeeper"
+            )
+        end
+    end
+
+    -- The last one thing: persist our vclock
+    ::set_vclockkeeper::
+    local vclock = box.info.vclock
+
+    -- WARNING: implicit yield
+    local ok, err = session:set_vclockkeeper(
+        replicaset_uuid, instance_uuid, vclock
+    )
+    fiber.testcancel()
+
+    if ok == nil then
+        return nil, SwitchoverError:new(err)
+    end
+
+    -- Hooray, instance is a legal vclockkeeper now.
+    vars.cache.is_vclockkeeper = true
+    vars.cache.is_leader = true
+    vars.cache.is_rw = true
+
+    log.info('Vclock persisted: %s',
+        json.encode(setmetatable(vclock, {_serialize = 'sequence'}))
+    )
+
+    return true
+end
+
+local function reconfigure_all(active_leaders)
     local confapplier = require('cartridge.confapplier')
     local all_roles = require('cartridge.roles').get_all_roles()
+
+::start_over::
+
+    local t1 = fiber.time()
+    -- WARNING: implicit yield
+    local ok, _ = constitute_oneself(active_leaders, {
+        timeout = vars.options.WAITLSN_TIMEOUT,
+    })
+    fiber.testcancel()
+    local t2 = fiber.time()
+
+    if not ok then
+        fiber.sleep(t1 + vars.options.WAITLSN_TIMEOUT - t2)
+        goto start_over
+    end
 
     -- WARNING: implicit yield
     -- The event may arrive while two-phase commit is in progress.
@@ -341,10 +485,12 @@ local function cfg(clusterwide_config)
 
     if failover_cfg.mode == 'disabled' then
         log.info('Failover disabled')
+        vars.consistency_needed = false
         first_appointments = _get_appointments_disabled_mode(topology_cfg)
 
     elseif failover_cfg.mode == 'eventual' then
         log.info('Eventual failover enabled')
+        vars.consistency_needed = false
         first_appointments = _get_appointments_eventual_mode(topology_cfg)
 
         vars.failover_fiber = fiber.new(failover_loop, {
@@ -356,6 +502,14 @@ local function cfg(clusterwide_config)
         vars.failover_fiber:name('cartridge.eventual-failover')
 
     elseif failover_cfg.mode == 'stateful' then
+        if topology_cfg.replicasets[box.info.cluster.uuid].all_rw then
+            -- Replicasets with all_rw flag imply that
+            -- consistent switchover isn't necessary
+            vars.consistency_needed = false
+        else
+            vars.consistency_needed = true
+        end
+
         if failover_cfg.state_provider == 'tarantool' then
             local params = assert(failover_cfg.tarantool_params)
             vars.client = stateboard_client.new({
@@ -369,6 +523,8 @@ local function cfg(clusterwide_config)
                 params.uri
             )
         elseif failover_cfg.state_provider == 'etcd2' then
+            -- TODO vclockkeeper handling in etcd2
+            vars.consistency_needed = false
             local params = assert(failover_cfg.etcd2_params)
             vars.client = etcd2_client.new({
                 endpoints = params.endpoints,
@@ -390,7 +546,7 @@ local function cfg(clusterwide_config)
             )
         end
 
-        -- WARNING: network yields
+        -- WARNING: implicit yield
         local appointments, err = _get_appointments_stateful_mode(vars.client, 0)
         if appointments == nil then
             log.warn('Failed to get first appointments: %s', err)
@@ -418,6 +574,21 @@ local function cfg(clusterwide_config)
     end
 
     accept_appointments(first_appointments)
+
+    local ok, err = constitute_oneself(vars.cache.active_leaders, {
+        timeout = vars.options.WAITLSN_TIMEOUT
+    })
+    if ok == nil then
+        log.warn("Error reaching consistency: %s", err)
+        if next(vars.schedule) == nil then
+            local task = fiber.new(reconfigure_all, vars.cache.active_leaders)
+            local id = task:id()
+            task:name('cartridge.failover.task')
+            vars.schedule[id] = task
+            log.info('Consistency not reached, another attempt scheduled (fiber %d)', id)
+        end
+    end
+
     box.cfg({
         read_only = not vars.cache.is_rw,
     })
@@ -432,6 +603,7 @@ end
 local function get_active_leaders()
     return vars.cache.active_leaders
 end
+
 --- Check current instance leadership.
 -- @function is_leader
 -- @local
@@ -446,6 +618,22 @@ end
 -- @treturn boolean true / false
 local function is_rw()
     return vars.cache.is_rw
+end
+
+--- Check if current instance has persisted his vclock.
+-- @function is_vclockkeeper
+-- @local
+-- @treturn boolean true / false
+local function is_vclockkeeper()
+    return vars.cache.is_vclockkeeper
+end
+
+--- Check if current configuration implies consistent switchover.
+-- @function consistency_needed
+-- @local
+-- @treturn boolean true / false
+local function consistency_needed()
+    return vars.consistency_needed
 end
 
 --- Get current stateful failover coordinator
@@ -479,8 +667,11 @@ end
 return {
     cfg = cfg,
     get_active_leaders = get_active_leaders,
-    is_leader = is_leader,
-    is_rw = is_rw,
     get_coordinator = get_coordinator,
     get_error = get_error,
+
+    consistency_needed = consistency_needed,
+    is_vclockkeeper = is_vclockkeeper,
+    is_leader = is_leader,
+    is_rw = is_rw,
 }